fix failover_critical.py alert on SCALE HA
